Move stream API routes to app router
